requirements.txt
twitchio
requests
python-dotenv
